RG-2542 Moved TableRow and TableCell to table-primitives.tsx
